--- typescript/codepipeline-build-deploy-github/.gitignore
@@ -1,10 +1,0 @@
-*.js
-!jest.config.js
-!*/trigger-build.js
-*.d.ts
-node_modules
-
-# CDK asset staging directory
-.cdk.staging
-cdk.out
-report.xml
