--- typescript/codepipeline-build-deploy-github/jest.config.js
@@ -1,14 +1,0 @@
-module.exports = {
-  testEnvironment: 'node',
-  roots: ['<rootDir>/test'],
-  testMatch: ['**/*.test.ts'],
-  transform: {
-    '^.+\\.tsx?$': 'ts-jest'
-  },
-  reporters: [
-    'default',
-    [ 'jest-junit', {
-      outputName: 'report.xml',
-    } ]
-  ]
-};
