--- typescript/codepipeline-build-deploy-github/lib/lambda/github_repo_creator/app/Dockerfile
@@ -1,7 +1,0 @@
-FROM public.ecr.aws/nginx/nginx:stable-alpine
-
-COPY ./site-assets/ /usr/share/nginx/html
-
-EXPOSE 80
-
-CMD ["nginx", "-g", "daemon off;"]
