--- typescript/codepipeline-build-deploy-github/lib/lambda/github_repo_creator/app/README.md
@@ -1,5 +1,0 @@
-# Simple Containerized Application
-
-## Overview
-
-This repository contains a simple application that will run in an Nginx container. The image will be built with Docker and stored in Amazon Elastic Container Registry (ECR).
