--- typescript/codepipeline-build-deploy-github/lib/lambda/github_repo_creator/app/appspec.yaml
@@ -1,9 +1,0 @@
-version: 0.0
-Resources:
-  - TargetService:
-      Type: AWS::ECS::Service
-      Properties:
-        TaskDefinition: "TASK_DEFINITION_ARN"
-        LoadBalancerInfo:
-          ContainerName: "web"
-          ContainerPort: "80"
