--- typescript/codepipeline-build-deploy-github/lib/lambda/github_repo_creator/app/site-assets/error.html
@@ -1,13 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-  <head>
-    <meta charset="UTF-8" />
-    <meta http-equiv="X-UA-Compatible" content="IE=edge" />
-    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <link rel="stylesheet" href="styles.css" />
-    <title>Error</title>
-  </head>
-  <body id="error">
-    <h1>Sorry, that page doesn't exist...</h1>
-  </body>
-</html>
