--- typescript/codepipeline-build-deploy-github/lib/lambda/github_repo_creator/app/site-assets/index.html
@@ -1,20 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-  <head>
-    <meta charset="UTF-8" />
-    <meta http-equiv="X-UA-Compatible" content="IE=edge" />
-    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <link rel="stylesheet" href="styles.css" />
-    <title>Homepage!</title>
-  </head>
-  <body>
-    <h1>Welcome to the Homepage!</h1>
-    <ul>
-      <li id="current"><a href="index.html">Home</a></li>
-      <li>|</li>
-      <li><a href="about.html">About</a></li>
-      <li>|</li>
-      <li><a href="contact.html">Contact</a></li>
-    </ul>
-  </body>
-</html>
