--- typescript/codepipeline-build-deploy-github/lib/lambda/github_repo_creator/app/taskdef.json
@@ -1,22 +1,0 @@
-{
-  "containerDefinitions": [
-    {
-      "name": "web",
-      "image": "REPOSITORY_URI:IMAGE_TAG",
-      "portMappings": [
-        {
-          "containerPort": 80,
-          "protocol": "tcp"
-        }
-      ],
-      "essential": true
-    }
-  ],
-  "taskRoleArn": "TASK_ROLE_ARN",
-  "executionRoleArn": "EXECUTION_ROLE_ARN",
-  "family": "codedeploy-sample",
-  "networkMode": "awsvpc",
-  "requiresCompatibilities": ["FARGATE"],
-  "cpu": "256",
-  "memory": "512"
-}
