--- typescript/codepipeline-build-deploy-github/package.json
@@ -1,28 +1,0 @@
-{
-  "name": "codepipeline-build-deploy",
-  "version": "0.1.0",
-  "bin": {
-    "codepipeline-build-deploy": "bin/codepipeline-build-deploy.js"
-  },
-  "scripts": {
-    "build": "tsc",
-    "watch": "tsc -w",
-    "test": "jest",
-    "cdk": "cdk"
-  },
-  "devDependencies": {
-    "@types/jest": "^29.4.0",
-    "@types/node": "18.14.6",
-    "aws-cdk": "2.72.1",
-    "jest": "^29.5.0",
-    "ts-jest": "^29.0.5",
-    "ts-node": "^10.9.1",
-    "typescript": "~5.1.6"
-  },
-  "dependencies": {
-    "aws-cdk-lib": "2.80.0",
-    "constructs": "^10.0.0",
-    "jest-junit": "^16.0.0",
-    "source-map-support": "^0.5.21"
-  }
-}
