+++ typescript/codepipeline-build-deploy/lib/lambda/github_repo_creator/app/Dockerfile
