+++ typescript/codepipeline-build-deploy/lib/lambda/github_repo_creator/app/site-assets/index.html
